Use bitcoin-core verify.py script
See: https://github.com/bitcoin/bitcoin/tree/master/contrib/verify-binaries

Currently, we due to the way the bitcoin core verify script parses the
version we must download multiple linux builds, then only use the the
target arch. Hopefully the verify script can be updated to handle single
platform (architecture:os) downloads, and if so in the future this can
be updated.

Switch to a two-stage build (like alpine), as we now download python
which bloats the image size.